Update login feature
add login feature for customer and check the user if the user is Admin,
the user can create, update, delete products and cateogries. Reset the
folder containing the file

+++ src/api/shop/ShopControllers.js
@@ -7,36 +7,45 @@ module.exports = {
   //Categories
   addCategory: async (req, res) => {
     const errors = {};
-    const { name, description } = req.body;
-    const cate = await Categories.findOne({ where: { name } });
-    if (cate) {
-      errors.cate = "Category already exists!";
-      return res.status(404).json(errors);
+    if (req.user.isAdmin) {
+      const { name, description } = req.body;
+      const cate = await Categories.findOne({ where: { name } });
+      if (cate) {
+        errors.cate = "Category already exists!";
+        return res.status(404).json(errors);
+      }
+      const newCate = { name, description };
+      Categories.create(newCate)
+        .then(cate => res.json(cate))
+        .catch(err => console.log(err));
+    } else {
+      errors.user = "You do not have access";
+      return res.status(403).json(errors);
     }
-    const newCate = { name, description };
-    Categories.create(newCate)
-      .then(cate => res.json(cate))
-      .catch(err => console.log(err));
   },
   updateCategory: async (req, res) => {
     const errors = {};
-    const id = req.params.cate_id;
-    const { name, description } = req.body;
+    if (req.user.isAdmin) {
+      const id = req.params.cate_id;
+      const { name, description } = req.body;
+      const isNameUnique = await Categories.findOne({
+        where: {
+          [Op.and]: [{ [Op.not]: { id } }, { name }]
+        }
+      });
 
-    const isNameUnique = await Categories.findOne({
-      where: {
-        [Op.and]: [{ [Op.not]: { id } }, { name }]
+      if (isNameUnique) {
+        errors.name = "Name already exists!";
+        return res.status(404).json(errors);
       }
-    });
-
-    if (isNameUnique) {
-      errors.name = "Name already exists!";
-      return res.status(404).json(errors);
+      const cate = await Categories.findById(id);
+      cate.update({ name, description }).then(cate => {
+        return res.json(cate);
+      });
+    } else {
+      errors.user = "You do not have access";
+      return res.status(403).json(errors);
     }
-    const cate = await Categories.findById(id);
-    cate.update({ name, description }).then(cate => {
-      return res.json(cate);
-    });
   },
   getListCate: async (req, res) => {
     const errors = {};
@@ -51,51 +60,69 @@ module.exports = {
     );
   },
   deleteCate: (req, res) => {
-    const id = req.params.cate_id;
-    Categories.destroy({ where: { id } }).then(() => {
-      return res.json({ success: true });
-    });
+    const errors = {};
+    if (req.user.isAdmin) {
+      const id = req.params.cate_id;
+      Categories.destroy({ where: { id } }).then(() => {
+        return res.json({ success: true });
+      });
+    } else {
+      errors.user = "You do not have access";
+      return res.status(403).json(errors);
+    }
   },
 
   // Products
   addProduct: (req, res) => {
-    const { cate_id, name, description, price } = req.body;
-    const image = req.file.filename;
-    Products.create({ cate_id, name, description, price, image })
-      .then(product => res.json(product))
-      .catch(err => console.log(err));
+    const errors = {};
+    if (req.user.isAdmin) {
+      const { cate_id, name, description, price } = req.body;
+      const image = req.file.filename;
+      Products.create({ cate_id, name, description, price, image })
+        .then(product => res.json(product))
+        .catch(err => console.log(err));
+    } else {
+      errors.user = "You do not have access";
+      return res.status(403).json(errors);
+    }
   },
   updateProduct: async (req, res) => {
-    const id = req.params.prod_id;
-    const product = await Products.findById(id).then(prod => prod);
-    const { cate_id, name, description, price } = req.body;
-    const { file } = req;
-    const updateProduct = { cate_id, name, description, price };
-    if (file != undefined) {
-      const filePath = `public/uploads/${product.image}`;
-      fs.unlinkSync(filePath);
-      Object.assign(
-        {},
-        updateProduct,
-        (updateProduct.image = req.file.filename)
-      );
+    const errors = {};
+    if (req.user.isAdmin) {
+      const id = req.params.prod_id;
+      const product = await Products.findById(id).then(prod => prod);
+      const { cate_id, name, description, price } = req.body;
+      const { file } = req;
+      const updateProduct = { cate_id, name, description, price };
+      if (file != undefined) {
+        const filePath = `public/uploads/${product.image}`;
+        fs.unlinkSync(filePath);
+        Object.assign(
+          {},
+          updateProduct,
+          (updateProduct.image = req.file.filename)
+        );
+        product
+          .update(updateProduct)
+          .then(prod => res.json(prod))
+          .catch(err => console.log(err));
+      }
       product
         .update(updateProduct)
         .then(prod => res.json(prod))
         .catch(err => console.log(err));
+    } else {
+      errors.user = "You do not have access";
+      return res.status(403).json(errors);
     }
-    product
-      .update(updateProduct)
-      .then(prod => res.json(prod))
-      .catch(err => console.log(err));
   },
   getListProducts: (req, res) => {
     const errors = {};
     Products.findAll()
       .then(products => {
-        if (!products) {
+        if (products.length <= 0) {
           errors.product = "Product not found!";
-          res.status(404).json(errors);
+          return res.status(404).json(errors);
         }
         res.json(products);
       })
@@ -130,19 +157,24 @@ module.exports = {
   },
   deleteProduct: async (req, res) => {
     const errors = {};
-    const id = req.params.prod_id;
-    const product = await Products.findById(id);
-    if (!product) {
-      errors.product = "Product not found";
-      return res.status(404).json(errors);
+    if (req.user.isAdmin) {
+      const id = req.params.prod_id;
+      const product = await Products.findById(id);
+      if (!product) {
+        errors.product = "Product not found";
+        return res.status(404).json(errors);
+      }
+      const filePath = `public/uploads/${product.image}`;
+      fs.unlink(filePath);
+      product
+        .destroy()
+        .then(() => {
+          return res.json({ success: "Delete success!" });
+        })
+        .catch(err => console.log(err));
+    } else {
+      errors.user = "You do not have access";
+      return res.status(403).json(errors);
     }
-    const filePath = `public/uploads/${product.image}`;
-    fs.unlink(filePath);
-    product
-      .destroy()
-      .then(() => {
-        return res.json({ success: "Delete success!" });
-      })
-      .catch(err => console.log(err));
   }
 };
